Refactored single bounty view
